Rework Dockerfile for smaller and fewer layers
This changes the main Dockerfile to no longer download files through Docker `ADD` directive and instead through `curl` on the same layer which removes it in the end. `ADD` creates layers that has to be downloaded even if the merged final image removes the files.

Several `ADD` and `RUN` lines were combined into other lines or removed to cut down on the total number of layers that have to be downloaded and extracted.

Some redundant `yum` lines were removed as well.