CUDA 9.2, Older CUDA, no AMD, TeXLive18
This branch adds CUDA 9.2 to the image, and also adds all the older versions of the CUDA toolkit back to 7.5.
CUDA is installed as normal and trimmed down a bit after, so the end result is less of a headache for us devs, but may include far more than we need, so the image size is larger.
Docs and examples are cut from each install

AMD SDK 3.0 is removed.
AMD pulled the SDK without notice and is only available through the graces of other users.
We have elected to remove the SDK from this version because of this choice and according to @peastman's memory '...the AMD headers was that compiling against the Nvidia ones produced slightly slower performance when using an AMD GPU on Windows...'. The small subset of users we expect this to effect also influenced our choice to remove the SDK.

TeXLive 2018 has been installed
As it says on the tin, but a new GLIBC was needed (see below)

GLIBC 2.14 has been added in parallel to the image
This version of GLIBC is now on the image because several core TeXLive 2018 packages needed it. This is a parallel install and does not disrupt the GLIBC 2.12 which is native to CentOS 6. The parallel GLIBC is not on the PATH or LD_LIBRARY_PATH, so should not be visible to any derivative programs.
Added to /opt/glibc-2.14